done. need fused_bias_dropout
